refactor(cli): extract repl command dispatch
